Fix Signed Value Overflow
Then converting from a BigInt, the first byte is the sign byte, which needs to be dropped before we deserialize the value, otherwise a large valid value `BigInt(1) << 255 - 1` which serializes to `0x007fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff` actually overflows (it's 33 bytes).

But there's an odd case where `BigInt(1) << 255 + 1` serializes to `0x008000000000000000000000000000000000000000000000000000000000000001`, which when we drop the first byte is `0x8000000000000000000000000000000000000000000000000000000000000001` which ends up actually being in the negative space (since the MSB is set), which means we've accidentally overflowed. This code adds a check for that and properly returns nil in this case.