Add Compliance Tests
This patch adds Compliance Tests. These are random odd-job Solidity files that we write that basically can be fuzz generated (not literally, but it also could be literally). We then run the code via `forge test` and capture their results to generate JSON files capturing the code run and the outputs of the tests. Then we re-run the same code in the EVM.swift and ensure the result matches what was generated in forge. This is a nice way to ensure behavior of the EVM.swift matches expectations.