Fix BigInt
BigInt expects that its serialized version has a first byte "0x00" during deserialization for the positive case. We were previously passing in the EthWord itself, which is 32-bytes wide, so usually it just started with a `0x00` that was ignored. But in this case, we had `0x40` as the most-significant byte, and _that_ was being consumed as the sign-byte, and thus ignored from the magnitude. Same issue as Issue #2, just on the other side. This patch adds the extra sign byte as expected.