Add Swift Syntax
This patch adds Swift Syntax. We add it to a new package `ABIGen`, so that it doesn't become a downstream dependency since it will likely end up as a separate CLI tool. The `ABI` code itself should live in Eth, but the CodeGen will be separate. We show that we _can_, in fact, produce Swift code using Apple-supporting tooling.