Fix Cool.swift Tests
This patch fixes `Cool.swift`'s tests. Specifically:

a) We fix a bug in `ABI.Function.outputTuple` that was incorrectly using the input tuple as the output tuple. D'oh
b) We fix it to use runtimeCode instead of creatoinCode. We switch to Solidity's terminology here to make that clearer.
c) We fix it so that we decode a `.tuple1(...)`, etc, and properly pull out basic parameters.

We will need to work on cases for arrays and structs, but this gets `Cool.swift` working correctly.