Add `Hex` Struct
This patch adds a `Hex` struct which simply wraps a `Data` object. The object is that this makes a lot of code significantly nicer without heavily modifying `Data`. E.g.

```
let a: Hex = "0x1122"
```

That will crash at compile-time if the hex value `0x1122` isn't valid. We make it quite easy to migrate to and from the underlying Data objects, e.g. via `a.data` and `Hex(Data([]))`, so the wrapper is light-weight.

Another key feature is that Hex will be properly describable, e.g. via:

```
print("Hello \(a)")
```

will print `Hello 0x1122` instead of `Hello [2 bytes]`, which is the behavior for `Data`.

Finally, we can make `Hex` decodable directly from Hex strings, thus if you encode or decode a struct with `Hex`, it will naturally encode and decode as Hex, instead of having to custom convert the value.

Note: I commented out the Geno tests as we'll need to rebase, and it's easier to make the change wholesale. This will be done prior to merge.