Error Handle in EVM
This patch adds the basic idea of how error handling could work in the EVM and how it could integrate into a generated Swift wrapper contract. Mostly, we add a function to `ABI.Function` to decode input (which will decode errors), and then we allow passing a set of functions to the `runQuery` function which will try reverts against those functions and raise those instead of a generic `.revert` when there's a match.