Add FFI Integration
This patch adds the ability to call into FFI "foreign-function interfaces," which are hooks back into the EVM's calling context environment. From the perspective of the EVM/Solidity code, this is the same as calling a standard Ethereum precompile (e.g. these functions: https://www.evm.codes/precompiled), but overall, this means we can allow hooks from EVM-VM code, e.g. to fetch offline price data. This lightly violates the concept of being pure, but overall, it's up to the discretion of the author to add hooks or not into code.

Note: we probably want FFIs to be async, which will require us to change `execVm` to become async, so waiting on feedback before making that change.