Add Console FFI
This patch adds the console FFI, allowing us to log messages from inside the EVM the same way that Forge/Foundry does.

The majority of this patch is simply handling Geno on some new corner cases (overloaded functions and void returns). These changes are a little hacky, but that's kinda the life of a code gen system.